requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-groq
pytesseract
Pillow
requests
beautifulsoup4
faiss-cpu
sentence-transformers
